refactor(cl): convert SearchTestCase to a mixin
