+++ cl/lib/test_helpers.py
@@ -25,13 +25,11 @@ from cl.search.constants import o_type_index_map
 from cl.search.docket_sources import DocketSources
 from cl.search.documents import DocketDocument
 from cl.search.factories import (
-    CitationWithParentsFactory,
     CourtFactory,
     DocketEntryWithParentsFactory,
     DocketFactory,
     OpinionClusterFactory,
     OpinionFactory,
-    OpinionsCitedWithParentsFactory,
     RECAPDocumentFactory,
 )
 from cl.search.models import (
@@ -783,217 +781,6 @@ audio_v4_fields.update(
         "meta": [],  # type: ignore
     }
 )
-
-
-class SearchTestCase(TestCase):
-    """Search test case factories"""
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.docket_1 = DocketFactory.create(
-            date_reargument_denied=datetime.date(2015, 8, 15),
-            court_id=cls.court_2.pk,
-            case_name_full="case name full docket 1",
-            date_argued=datetime.date(2015, 8, 16),
-            case_name="case name docket 1",
-            case_name_short="case name short docket 1",
-            docket_number="docket number 1 005",
-            slug="case-name",
-            pacer_case_id="666666",
-            blocked=False,
-            source=0,
-        )
-        cls.docket_2 = DocketFactory.create(
-            date_reargument_denied=datetime.date(2015, 8, 15),
-            court_id=cls.court_2.pk,
-            case_name_full="case name full docket 2",
-            date_argued=datetime.date(2015, 8, 15),
-            case_name="case name docket 2",
-            case_name_short="case name short docket 2",
-            docket_number="docket number 2",
-            slug="case-name",
-            blocked=False,
-            source=0,
-        )
-        cls.docket_3 = DocketFactory.create(
-            date_reargument_denied=datetime.date(2015, 8, 15),
-            court_id=cls.court_1.pk,
-            case_name_full="case name full docket 3",
-            date_argued=datetime.date(2015, 8, 14),
-            case_name="case name docket 3",
-            case_name_short="case name short docket 3",
-            docket_number="docket number 3",
-            slug="case-name",
-            blocked=False,
-            source=0,
-        )
-
-        cls.opinion_cluster_1 = OpinionClusterFactory.create(
-            case_name_full="Paul Debbas v. Franklin",
-            case_name_short="Debbas",
-            syllabus="some rando syllabus",
-            date_filed=datetime.date(2015, 8, 14),
-            procedural_history="some rando history",
-            source="C",
-            case_name="Debbas v. Franklin",
-            attorneys="a bunch of crooks!",
-            slug="case-name-cluster",
-            precedential_status="Errata",
-            citation_count=4,
-            docket=cls.docket_1,
-        )
-        cls.opinion_cluster_1.panel.add(cls.person_2)
-
-        cls.opinion_cluster_2 = OpinionClusterFactory.create(
-            case_name_full="Harvey Howard v. Antonin Honda",
-            case_name_short="Howard",
-            syllabus="some rando syllabus",
-            date_filed=datetime.date(1895, 6, 9),
-            procedural_history="some rando history",
-            source="C",
-            judges="David",
-            case_name="Howard v. Honda",
-            attorneys="a bunch of crooks!",
-            slug="case-name-cluster",
-            precedential_status="Published",
-            citation_count=6,
-            scdb_votes_minority=3,
-            scdb_votes_majority=6,
-            nature_of_suit="copyright",
-            docket=cls.docket_2,
-        )
-
-        cls.opinion_cluster_3 = OpinionClusterFactory.create(
-            case_name_full="Reference to Lissner v. Saad",
-            case_name_short="case name short cluster 3",
-            syllabus="some rando syllabus",
-            date_filed=datetime.date(2015, 8, 15),
-            procedural_history="some rando history",
-            source="C",
-            case_name="case name cluster 3",
-            attorneys="a bunch of crooks!",
-            slug="case-name-cluster",
-            precedential_status="Published",
-            citation_count=8,
-            docket=cls.docket_3,
-        )
-
-        cls.citation_1 = CitationWithParentsFactory.create(
-            volume=33,
-            reporter="state",
-            page="1",
-            type=1,
-            cluster=cls.opinion_cluster_1,
-        )
-        cls.citation_2 = CitationWithParentsFactory.create(
-            volume=22,
-            reporter="AL",
-            page="339",
-            type=8,
-            cluster=cls.opinion_cluster_2,
-        )
-        cls.citation_3 = CitationWithParentsFactory.create(
-            volume=33,
-            reporter="state",
-            page="1",
-            type=1,
-            cluster=cls.opinion_cluster_2,
-        )
-        cls.citation_4 = CitationWithParentsFactory.create(
-            volume=1,
-            reporter="Yeates",
-            page="1",
-            type=5,
-            cluster=cls.opinion_cluster_2,
-        )
-        cls.citation_5 = CitationWithParentsFactory.create(
-            volume=56,
-            reporter="F.2d",
-            page="9",
-            type=1,
-            cluster=cls.opinion_cluster_2,
-        )
-        cls.citation_5 = CitationWithParentsFactory.create(
-            volume=56,
-            reporter="F.2d",
-            page="11",
-            type=1,
-            cluster=cls.opinion_cluster_3,
-        )
-        cls.opinion_1 = OpinionFactory.create(
-            extracted_by_ocr=False,
-            author=cls.person_2,
-            plain_text="my plain text secret word for queries",
-            cluster=cls.opinion_cluster_1,
-            local_path="test/search/opinion_doc.doc",
-            per_curiam=False,
-            type="020lead",
-        )
-        cls.opinion_2 = OpinionFactory.create(
-            extracted_by_ocr=False,
-            author=cls.person_2,
-            plain_text="my plain text secret word for queries",
-            cluster=cls.opinion_cluster_2,
-            html_with_citations='yadda yadda <span class="star-pagination">*9</span> this is page 9 <span class="star-pagination">*10</span> this is content on page 10 can we link to it...',
-            local_path="test/search/opinion_pdf_image_based.pdf",
-            per_curiam=False,
-            type="010combined",
-        )
-        cls.opinion_2.joined_by.add(cls.person_2)
-
-        cls.opinion_3 = OpinionFactory.create(
-            extracted_by_ocr=False,
-            author=cls.person_2,
-            plain_text="my plain text secret word for queries 1 Yeates 1",
-            cluster=cls.opinion_cluster_3,
-            local_path="test/search/opinion_pdf_text_based.pdf",
-            per_curiam=False,
-            type="010combined",
-        )
-        cls.opinion_4 = OpinionFactory.create(
-            extracted_by_ocr=False,
-            author=cls.person_2,
-            plain_text="my plain text secret word for queries",
-            cluster=cls.opinion_cluster_1,
-            local_path="test/search/opinion_html.html",
-            per_curiam=False,
-            type="010combined",
-        )
-        cls.opinion_5 = OpinionFactory.create(
-            extracted_by_ocr=False,
-            author=cls.person_2,
-            plain_text="my plain text secret word for queries",
-            cluster=cls.opinion_cluster_1,
-            local_path="test/search/opinion_wpd.wpd",
-            per_curiam=False,
-            type="010combined",
-        )
-        cls.opinion_6 = OpinionFactory.create(
-            extracted_by_ocr=False,
-            author=cls.person_2,
-            plain_text="my plain text secret word for queries",
-            cluster=cls.opinion_cluster_1,
-            local_path="test/search/opinion_text.txt",
-            per_curiam=False,
-            type="010combined",
-        )
-        cls.opinion_cited_1 = OpinionsCitedWithParentsFactory.create(
-            cited_opinion=cls.opinion_2,
-            citing_opinion=cls.opinion_1,
-        )
-        cls.opinion_cited_2 = OpinionsCitedWithParentsFactory.create(
-            cited_opinion=cls.opinion_3,
-            citing_opinion=cls.opinion_1,
-        )
-        cls.opinion_cited_3 = OpinionsCitedWithParentsFactory.create(
-            cited_opinion=cls.opinion_3,
-            citing_opinion=cls.opinion_2,
-        )
-        cls.opinion_cited_4 = OpinionsCitedWithParentsFactory.create(
-            cited_opinion=cls.opinion_1,
-            citing_opinion=cls.opinion_3,
-        )
-        super().setUpTestData()
 
 
 class RECAPSearchTestCase(TestCase):
