+++ cl/opinion_page/tests.py
@@ -31,7 +31,6 @@ from cl.lib.models import THUMBNAIL_STATUSES
 from cl.lib.redis_utils import get_redis_interface
 from cl.lib.storage import clobbering_get_name
 from cl.lib.test_helpers import (
-    SearchTestCase,
     SimpleUserDataMixin,
     SitemapTest,
 )
@@ -88,7 +87,7 @@ from cl.search.models import (
 )
 from cl.sitemaps_infinite.sitemap_generator import generate_urls_chunk
 from cl.tests.cases import ESIndexTestCase, TestCase
-from cl.tests.mixins import CourtMixin, PeopleMixin
+from cl.tests.mixins import CourtMixin, PeopleMixin, SearchMixin
 from cl.tests.providers import fake
 from cl.users.factories import UserFactory, UserProfileWithParentsFactory
 
@@ -120,8 +119,8 @@ class SimpleLoadTest(TestCase):
 class OpinionPageLoadTest(
     CourtMixin,
     PeopleMixin,
+    SearchMixin,
     ESIndexTestCase,
-    SearchTestCase,
     TestCase,
 ):
     @classmethod
